Refs #11. refactor cncs inel
